moving chunks for descriptors, fixing descriptor bugs, and zscore for prediction
