+++ src/main/java/edu/unc/ceccr/chembench/workflows/descriptors/DescriptorCDK.java
@@ -7,12 +7,7 @@ import edu.unc.ceccr.chembench.utilities.RunExternalProgram;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.FileReader;
-import java.nio.file.Path;
-import java.nio.file.Paths;
+import java.io.*;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
@@ -53,7 +48,101 @@ public class DescriptorCDK implements DescriptorSet{
         convertCdkToX(outfileWithEnding, workingDir + "/Descriptors/");
     }
 
-    public void convertCdkToX(String cdkOutputFile, String workingDir) throws Exception {
+    @Override
+    public void readDescriptors(String xFile, List<String> descriptorNames,
+                                        List<Descriptors> descriptorValueMatrix) throws Exception {
+        xFile += getFileEnding();
+        String cdkXFile = xFile + ".x";
+
+        //generate x file for cdk if it doesn't exist already
+        if (!new File(cdkXFile).isFile()) {
+            File cdkFile = new File(xFile);
+            if (!cdkFile.exists() || cdkFile.length() == 0) {
+                logger.error(xFile + ": cdk file not found");
+                throw new Exception("Could not read cdk file descriptors for cdk to x file operation: " + xFile + "\n");
+            }
+            convertCdkToX(xFile, cdkFile.getParent());
+        }
+
+        readDescriptorFile(cdkXFile, descriptorNames, descriptorValueMatrix);
+    }
+
+    @Override
+    public void readDescriptorsChunks(String xFile, List<String> descriptorNames,
+                                      List<Descriptors> descriptorValueMatrix) throws Exception {
+
+        readDescriptorFile(xFile, descriptorNames, descriptorValueMatrix);
+    }
+
+    @Override
+    public String splitFile(String workingDir, String descriptorsFile) throws Exception {
+        descriptorsFile += getFileEnding();
+        convertCdkToX(workingDir + descriptorsFile, workingDir);
+        descriptorsFile += ".x";
+
+        File file = new File(workingDir + descriptorsFile);
+        if (!file.exists() || file.length() == 0) {
+            throw new Exception("Could not read UPLOADED descriptors.\n");
+        }
+        FileReader fin = new FileReader(file);
+        BufferedReader br = new BufferedReader(fin);
+
+        int currentFile = 0;
+        int moleculesInCurrentFile = 0;
+        BufferedWriter outFilePart =
+                new BufferedWriter(new FileWriter(workingDir + descriptorsFile + "_" + currentFile));
+
+        // don't bother changing the numbers to reflect #compounds in file
+        // part, it doesn't matter
+        String header = br.readLine() + "\n";
+        String descriptorNames = br.readLine() + "\n";
+        outFilePart.write(header);
+        outFilePart.write(descriptorNames);
+
+        String line;
+        while ((line = br.readLine()) != null) {
+            outFilePart.write(line + "\n");
+
+            moleculesInCurrentFile++;
+            if (moleculesInCurrentFile == compoundsPerChunk) {
+                outFilePart.close();
+                moleculesInCurrentFile = 0;
+                currentFile++;
+                outFilePart = new BufferedWriter(new FileWriter(workingDir + descriptorsFile + "_" + currentFile));
+                outFilePart.write(header);
+                outFilePart.write(descriptorNames);
+            }
+        }
+        br.close();
+        outFilePart.write("\n");
+        outFilePart.close();
+
+        return descriptorsFile;
+    }
+
+    @Override
+    public String checkDescriptors(String cdkOutputFile) throws Exception {
+        // check if CDK file exists, then look for NA's.
+        String errors = "";
+
+        File file = new File(cdkOutputFile + getFileEnding());
+        if (!file.exists() || file.length() == 0) {
+            errors = "Could not read descriptor file.\n";
+        } else {
+            BufferedReader br = new BufferedReader(new FileReader(file));
+            String line = "";
+            while ((line = br.readLine()) != null) {
+                if (line.contains("NA")) {
+                    String compoundName = line.split("\\s+")[0];
+                    errors += ("Error: compound named '" + compoundName + "' has some invalid descriptors. \n");
+                }
+            }
+            br.close();
+        }
+        return errors;
+    }
+
+    private void convertCdkToX(String cdkOutputFile, String workingDir) throws Exception {
         String cmd = "python " + Constants.CECCR_BASE_PATH + Constants.SCRIPTS_PATH + "cdkToX.py " + cdkOutputFile +
                 " " + cdkOutputFile + ".x";
         RunExternalProgram.runCommandAndLogOutput(cmd, workingDir, "cdkToX.py");
@@ -62,28 +151,14 @@ public class DescriptorCDK implements DescriptorSet{
         // 'em.
     }
 
-    @Override
-    public void readDescriptors(String xFile, List<String> descriptorNames,
-                                        List<Descriptors> descriptorValueMatrix) throws Exception {
-
-        String cdkXFile = xFile + ".x";
-        File file = new File(cdkXFile);
-
-        //generate x file for cdk if it doesn't exist already
-        if (!file.exists()) {
-            File sdfFile = new File(xFile);
-            if (!sdfFile.exists() || sdfFile.length() == 0) {
-                logger.error(xFile + ": sdf file not found");
-                throw new Exception("Could not read sdf file descriptors for cdk to x file operation: " + xFile + "\n");
-            }
-             convertCdkToX(xFile, sdfFile.getParent());
-            file = new File(cdkXFile);
-        }
-
+    private void readDescriptorFile(String xFile, List<String> descriptorNames,
+                                    List<Descriptors> descriptorValueMatrix) throws Exception{
+        //read generated x file
+        File file = new File(xFile);
         logger.debug("Trying to read uploaded descriptors");
         if (!file.exists() || file.length() == 0) {
-            logger.error(cdkXFile + ": cdkXFile not found");
-            throw new Exception("Could not read X file descriptors: " + cdkXFile + "\n");
+            logger.error(xFile + ": xFile not found");
+            throw new Exception("Could not read X file descriptors: " + xFile + "\n");
         }
 
         try {
@@ -122,27 +197,5 @@ public class DescriptorCDK implements DescriptorSet{
         } catch (FileNotFoundException e) {
             logger.error(file + ": File not found");
         }
-    }
-
-    @Override
-    public String checkDescriptors(String cdkOutputFile) throws Exception {
-        // check if CDK file exists, then look for NA's.
-        String errors = "";
-
-        File file = new File(cdkOutputFile + getFileEnding());
-        if (!file.exists() || file.length() == 0) {
-            errors = "Could not read descriptor file.\n";
-        } else {
-            BufferedReader br = new BufferedReader(new FileReader(file));
-            String line = "";
-            while ((line = br.readLine()) != null) {
-                if (line.contains("NA")) {
-                    String compoundName = line.split("\\s+")[0];
-                    errors += ("Error: compound named '" + compoundName + "' has some invalid descriptors. \n");
-                }
-            }
-            br.close();
-        }
-        return errors;
     }
 }
