+++ src/main/java/edu/unc/ceccr/chembench/workflows/descriptors/DescriptorMaccs.java
@@ -6,9 +6,7 @@ import edu.unc.ceccr.chembench.utilities.RunExternalProgram;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileReader;
+import java.io.*;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
@@ -45,10 +43,73 @@ public class DescriptorMaccs implements DescriptorSet{
     public void readDescriptors(String maccsOutputFile, List<String> descriptorNames,
                    List<Descriptors> descriptorValueMatrix) throws Exception {
         // generate with "maccs.sh infile.sdf outfile.maccs"
+        maccsOutputFile +=getFileEnding();
+        readDescriptorFile (maccsOutputFile, descriptorNames, descriptorValueMatrix);
+    }
 
+    @Override
+    public void readDescriptorsChunks(String outputFile, List<String> descriptorNames,
+                                      List<Descriptors> descriptorValueMatrix) throws Exception {
+        readDescriptorFile (outputFile, descriptorNames, descriptorValueMatrix);
+    }
+
+    @Override
+    public String splitFile(String workingDir, String descriptorsFile) throws Exception {
+        descriptorsFile += getFileEnding();
+
+        File file = new File(workingDir + descriptorsFile);
+        if (!file.exists() || file.length() == 0) {
+            throw new Exception("Could not read Maccs descriptors.\n");
+        }
+        FileReader fin = new FileReader(file);
+        BufferedReader br = new BufferedReader(fin);
+
+        String header = ""; // stores everything up to where descriptors
+        // begin.
+        int currentFile = 0;
+        int moleculesInCurrentFile = 0;
+        BufferedWriter outFilePart =
+                new BufferedWriter(new FileWriter(workingDir + descriptorsFile + "_" + currentFile));
+
+        header = br.readLine() + "\n";
+        outFilePart.write(header);
+
+        String line;
+        while ((line = br.readLine()) != null) {
+            outFilePart.write(line + "\n");
+
+            moleculesInCurrentFile++;
+            if (moleculesInCurrentFile == compoundsPerChunk) {
+                outFilePart.close();
+                moleculesInCurrentFile = 0;
+                currentFile++;
+                outFilePart = new BufferedWriter(new FileWriter(workingDir + descriptorsFile + "_" + currentFile));
+                outFilePart.write(header);
+            }
+        }
+        br.close();
+        outFilePart.close();
+
+        return descriptorsFile;
+    }
+
+    @Override
+    public String checkDescriptors(String maccsOutputFile) throws Exception {
+        // right now this doesn't check anything. The MACCS keys never seem to
+        // cause issues.
+        String errors = "";
+
+        File file = new File(maccsOutputFile + getFileEnding());
+        if (!file.exists() || file.length() == 0) {
+            errors = "Could not read descriptor file.\n";
+        }
+        return errors;
+    }
+
+    private void readDescriptorFile (String outputFile, List<String> descriptorNames, List<Descriptors>
+            descriptorValueMatrix) throws Exception{
         logger.debug("reading Maccs Descriptors");
-
-        File file = new File(maccsOutputFile);
+        File file = new File(outputFile);
         if (!file.exists() || file.length() == 0) {
             throw new Exception("Could not read MACCS keys.\n");
         }
@@ -89,17 +150,5 @@ public class DescriptorMaccs implements DescriptorSet{
         for (int i = 0; i < Constants.NUM_MACCS_KEYS; i++) {
             descriptorNames.add((new Integer(i)).toString());
         }
-    }
-    @Override
-    public String checkDescriptors(String maccsOutputFile) throws Exception {
-        // right now this doesn't check anything. The MACCS keys never seem to
-        // cause issues.
-        String errors = "";
-
-        File file = new File(maccsOutputFile + getFileEnding());
-        if (!file.exists() || file.length() == 0) {
-            errors = "Could not read descriptor file.\n";
-        }
-        return errors;
     }
 }
