+++ src/main/java/edu/unc/ceccr/chembench/workflows/descriptors/WriteDescriptors.java
@@ -437,6 +437,34 @@ public class WriteDescriptors {
         br.close();
     }
 
+    private static void addZeroToOutput(List<Descriptors> descriptorMatrix,
+                                        StringBuffer descriptorNameStringBuffer,
+                                        String predictorDescriptorNameString){
+        logger.info("Adding 0's to the dataset descriptors");
+        String[] datasetDescriptorNames = descriptorNameStringBuffer.toString().split("\\s+");
+        String[] predictorDescriptorNames = predictorDescriptorNameString.split("\\s+");
+
+        //assumption that all descriptors in output/dataset is in predictor
+        //because output/dataset is the intersection of itself and predictor
+
+        //j keeps track of the dataset while i keeps track of the predictor
+        int j = 0;
+        for (Descriptors descriptors: descriptorMatrix) {
+            List<Double> descriptorValues= new ArrayList<>();
+            for (int i = 0; i < predictorDescriptorNames.length; i++) {
+                if ((j>=datasetDescriptorNames.length) ||
+                        (!datasetDescriptorNames[j].equals(predictorDescriptorNames[i]))) {
+                    descriptorValues.add(0.0);
+                } else {
+                    descriptorValues.add(descriptors.getDescriptorValueAtIndex(j++));
+                }
+            }
+            descriptors.setDescriptorValues(descriptorValues);
+        }
+        descriptorNameStringBuffer.setLength(0);
+        descriptorNameStringBuffer.append(predictorDescriptorNameString);
+    }
+
     public static void writeModelingXFile(List<String> compoundNames, List<Descriptors> descriptorMatrix,
                                           List<String> descriptorNames, String xFilePath, String scalingType,
                                           String stdDevCutoff, String correlationCutoff) throws Exception {
@@ -552,34 +580,6 @@ public class WriteDescriptors {
         } catch (IOException e) {
             throw new RuntimeException("Couldn't write modeling X file", e);
         }
-    }
-
-    public static void addZeroToOutput(List<Descriptors> descriptorMatrix,
-                                       StringBuffer descriptorNameStringBuffer,
-                                       String predictorDescriptorNameString){
-        logger.info("Adding 0's to the dataset descriptors");
-        String[] datasetDescriptorNames = descriptorNameStringBuffer.toString().split("\\s+");
-        String[] predictorDescriptorNames = predictorDescriptorNameString.split("\\s+");
-
-        //assumption that all descriptors in output/dataset is in predictor
-        //because output/dataset is the intersection of itself and predictor
-
-        //j keeps track of the dataset while i keeps track of the predictor
-        int j = 0;
-        for (Descriptors descriptors: descriptorMatrix) {
-            List<Double> descriptorValues= new ArrayList<>();
-            for (int i = 0; i < predictorDescriptorNames.length; i++) {
-                if ((j>=datasetDescriptorNames.length) ||
-                        (!datasetDescriptorNames[j].equals(predictorDescriptorNames[i]))) {
-                    descriptorValues.add(0.0);
-                } else {
-                    descriptorValues.add(descriptors.getDescriptorValueAtIndex(j++));
-                }
-            }
-            descriptors.setDescriptorValues(descriptorValues);
-        }
-        descriptorNameStringBuffer.setLength(0);
-        descriptorNameStringBuffer.append(predictorDescriptorNameString);
     }
 
     public static void writePredictionXFile(List<String> compoundNames, List<Descriptors> descriptorMatrix,
